refactor: index.js separado nos arquivos app.js e server.js para poder implementar os testes mais facilmente
