+++ backend/src/server/app.js
@@ -1,0 +1,13 @@
+const express = require('express');
+const cors = require('cors');
+
+const taskRoutes = require('../routes/taskRoutes');
+
+const app = express();
+
+app.use(cors());
+app.use(express.json());
+
+app.use('/tasks', taskRoutes);
+
+module.exports = app;
