+++ backend/src/server/server.js
@@ -1,0 +1,5 @@
+const app = require('./app');
+
+const PORT = 4000;
+
+app.listen(PORT, () => console.log(`Rodando na porta: ${PORT}`));
